% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classMethodDefine.R
\name{Ops.altWrapper}
\alias{Ops.altWrapper}
\alias{range,altWrapper-method}
\alias{Math.altWrapper}
\alias{range.altWrapper}
\title{altWrapper group generic functions}
\usage{
\method{Ops}{altWrapper}(e1, e2)

\method{Math}{altWrapper}(x, ...)

\method{range}{altWrapper}(..., na.rm = FALSE, finite = FALSE)
}
\arguments{
\item{e1}{left operand}

\item{e2}{right operand}

\item{x}{argument}

\item{...}{additional arguments}

\item{na.rm}{logical, indicating if NA's should be omitted.}

\item{finite}{logical, indicating if all non-finite elements should be omitted.}
}
\value{
Please see the document of R's native functions to find the return value
}
\description{
The usage of the group generic functions is the same as R's native functions.
Please see R's document for help.
}
