% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDefine.R
\name{newAltWrapper}
\alias{newAltWrapper}
\title{Make an ALTREP S3 or S4 class}
\usage{
newAltWrapper(x, S3Class = TRUE, S4Class, forceSet = FALSE)
}
\arguments{
\item{x}{an ALTREP object.}

\item{S3Class}{logical, whether the result is an S3 object}

\item{S4Class}{logical, whether the result is an S4 object, if both
S3Class and S4Class are TRUE, the package will use S4Class}

\item{forceSet}{logical, force R to set the class attribute for the variable x, it
violates R's copy-on-change philosophy and cause the variable x
outside of the function also changes. It is useful only when the normal way to create
the class object does not work. The argument only works when \code{S3Class = TRUE}}
}
\description{
This function will create a wrapper for an ALTREP object.
Most functions in \code{?groupGeneric} have been overloaded and support
ALTREP object.
}
